requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
Authlib==1.2.0
beautifulsoup4==4.12.2
boto3==1.26.130
botocore==1.29.130
bpemb==0.3.4
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
conllu==4.5.2
contourpy==1.0.7
cryptography==40.0.2
cssselect==1.2.0
cycler==0.11.0
datasets==2.10.1
decorator==5.1.1
Deprecated==1.2.13
dill==0.3.6
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.12.0
flair==0.12
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.5.0
ftfy==6.1.1
future==0.18.3
gdown==4.4.0
gensim==4.3.1
greenlet==2.0.2
huggingface-hub==0.10.0
hyperopt==0.2.7
idna==3.4
itsdangerous==2.1.2
Janome==0.4.2
jieba3k==0.35.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
kiwisolver==1.4.4
langdetect==1.0.9
lxml==4.9.2
MarkupSafe==2.1.2
matplotlib==3.7.1
more-itertools==9.1.0
mpld3==0.3
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
networkx==3.1
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.24.3
odict==1.9.0
openai==0.27.6
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
plumber==1.7
pptree==3.1
protobuf==3.20.2
py4j==0.10.9.7
pyarrow==12.0.0
pycparser==2.21
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytorch_revgrad==0.2.0
pytz==2023.3
PyYAML==6.0
regex==2023.5.5
requests==2.28.2
requests-file==1.5.1
responses==0.18.0
s3transfer==0.6.1
scikit-learn==1.2.2
scipy==1.10.1
segtok==1.5.11
sentencepiece==0.1.99
sgmllib3k==1.0.0
six==1.16.0
smart-open==6.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.12
sqlitedict==2.1.0
sympy==1.11.1
tabulate==0.9.0
threadpoolctl==3.1.0
tiktoken==0.4.0
tinysegmenter==0.3
tldextract==3.4.1
tokenizers==0.13.3
torch==2.0.1
tqdm==4.65.0
transformer-smaller-training-vocab==0.2.3
transformers==4.25.1
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
validators==0.20.0
wcwidth==0.2.6
weaviate-client==3.18.0
Werkzeug==2.3.4
Wikipedia-API==0.5.8
wrapt==1.15.0
xxhash==3.2.0
yarl==1.9.2
zope.component==6.0
zope.deferredimport==4.4
zope.deprecation==5.0
zope.event==4.6
zope.hookable==5.4
zope.interface==6.0
zope.lifecycleevent==4.4
zope.proxy==5.0.0
